feat: generic hints for configs
