+++ codex-rs/features/src/feature_configs.rs
@@ -1,4 +1,3 @@
-use crate::FeatureConfig;
 use schemars::JsonSchema;
 use serde::Deserialize;
 use serde::Serialize;
@@ -6,8 +5,6 @@ use serde::Serialize;
 #[derive(Serialize, Deserialize, Debug, Clone, Default, PartialEq, Eq, JsonSchema)]
 #[serde(deny_unknown_fields)]
 pub struct MultiAgentV2ConfigToml {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub enabled: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[schemars(range(min = 1))]
     pub max_concurrent_threads_per_session: Option<usize>,
@@ -26,31 +23,9 @@ pub struct MultiAgentV2ConfigToml {
     pub hide_spawn_agent_metadata: Option<bool>,
 }
 
-impl FeatureConfig for MultiAgentV2ConfigToml {
-    fn enabled(&self) -> Option<bool> {
-        self.enabled
-    }
-
-    fn set_enabled(&mut self, enabled: bool) {
-        self.enabled = Some(enabled);
-    }
-}
-
 #[derive(Serialize, Deserialize, Debug, Clone, Default, PartialEq, Eq, JsonSchema)]
 #[serde(deny_unknown_fields)]
 pub struct AppsMcpPathOverrideConfigToml {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub enabled: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
     pub path: Option<String>,
-}
-
-impl FeatureConfig for AppsMcpPathOverrideConfigToml {
-    fn enabled(&self) -> Option<bool> {
-        self.enabled.or(self.path.as_ref().map(|_| true))
-    }
-
-    fn set_enabled(&mut self, enabled: bool) {
-        self.enabled = Some(enabled);
-    }
 }
